refactor(common): interface 모듈 분리
