+++ src/main/java/com/yeohangttukttak/api/domain/file/entity/File.java
@@ -1,6 +1,6 @@
 package com.yeohangttukttak.api.domain.file.entity;
 
-import com.yeohangttukttak.api.domain.Attachable;
+import com.yeohangttukttak.api.global.interfaces.Attachable;
 import com.yeohangttukttak.api.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.Builder;
