+++ src/main/java/com/yeohangttukttak/api/global/interfaces/ValueBasedEnum.java
@@ -1,4 +1,4 @@
-package com.yeohangttukttak.api.domain;
+package com.yeohangttukttak.api.global.interfaces;
 
 public interface ValueBasedEnum {
     String getValue();
